Create a map-builder UI with Cytoscape
Building map JSON looked a bit tedious. I made this a separate directory outside of `src` and made it intentionally barren so we aren't tempted to go crazy with UI and accidentally make a D&D game in javascript.

I found a free drag-and-drop graph builder. It's designed for graph theory research, so it if very powerful, but for us it's just helpful.

There is also an extension which improves edge creation, but I would need webpack for it. Instead, I copy-pasted the code into a `vendor` directory.